requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
pydantic
sentence-transformers
rank-bm25
numpy
scikit-learn
tiktoken
langchain
langchain-openai
langchain-community
flask
openai>=1.0.0
sentence-transformers>=2.2.0
rank_bm25>=0.2.1
transformers>=4.30.0
python-dotenv>=0.19.0
requests
beautifulsoup4
pandas
tqdm
lightrag
graphrag
ragas
datasets
